feat: separate automation composables
